refactor agent services
