+++ server/agent/agentic_execution/tests/test_services.py
@@ -109,7 +109,7 @@ class TestStart:
 
     def test_attaches_files_as_conversation_files(self, service, agent, user):
         with patch("agent.agentic_execution.services.settings.FILE_PARSER_CLASSES", SUPPORTED_EXTENSIONS), \
-             patch("agent.agentic_execution.services.FileService.process", return_value="extracted"):
+             patch("agent.agentic_execution.services.FileParsingService.process", return_value="extracted"):
             execution = service.start(
                 agent=agent, user=user, execution_key="dummy", validated_input={}, uploaded_files=[_make_file()]
             )
@@ -123,7 +123,7 @@ class TestStart:
 
     def test_detects_image_file_category(self, service, agent, user):
         with patch("agent.agentic_execution.services.settings.FILE_PARSER_CLASSES", {(".jpg",): "some.Class"}), \
-             patch("agent.agentic_execution.services.FileService.process", return_value=""):
+             patch("agent.agentic_execution.services.FileParsingService.process", return_value=""):
             execution = service.start(
                 agent=agent, user=user, execution_key="dummy", validated_input={}, uploaded_files=[_make_file("photo.jpg", content_type="image/jpeg")]
             )
@@ -139,7 +139,7 @@ class TestStart:
     def test_creates_file_message_for_each_uploaded_file(self, service, agent, user):
         files = [_make_file("doc.pdf"), _make_file("sheet.txt")]
         with patch("agent.agentic_execution.services.settings.FILE_PARSER_CLASSES", SUPPORTED_EXTENSIONS), \
-             patch("agent.agentic_execution.services.FileService.process", return_value=""):
+             patch("agent.agentic_execution.services.FileParsingService.process", return_value=""):
             execution = service.start(
                 agent=agent, user=user, execution_key="dummy", validated_input={}, uploaded_files=files
             )
@@ -151,7 +151,7 @@ class TestStart:
 
     def test_file_message_has_correct_name_and_type(self, service, agent, user):
         with patch("agent.agentic_execution.services.settings.FILE_PARSER_CLASSES", SUPPORTED_EXTENSIONS), \
-             patch("agent.agentic_execution.services.FileService.process", return_value=""):
+             patch("agent.agentic_execution.services.FileParsingService.process", return_value=""):
             execution = service.start(
                 agent=agent, user=user, execution_key="dummy", validated_input={}, uploaded_files=[_make_file("report.pdf")]
             )
