+++ server/agent/services/conversation_file_message_service.py
@@ -5,17 +5,8 @@ from agent.models.message import Message
 
 
 class ConversationFileMessageService:
-    """Creates the FILE message that surfaces an attached file in the conversation history."""
-
-    def create_for_file(self, conversation_file: ConversationFile) -> Message:
-        """Create a FILE message for the given ConversationFile.
-
-        Args:
-            conversation_file: The already-persisted ConversationFile to create a message for.
-
-        Returns:
-            The newly created Message record.
-        """
+    @staticmethod
+    def create_for_file(conversation_file: ConversationFile) -> Message:
         file_name = conversation_file.file.name
         return Message.objects.create(
             conversation_id=conversation_file.conversation_id,
